# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Parabellum POS API",
    "description": "Spécification complète et interactive des API du système de point de vente (POS) et de commande en ligne Gourmet. Supporte l'authentification par session, les webhooks tiers et les protocoles matériels réseau.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement local"
    },
    {
      "url": "https://www.parabellumpos.online",
      "description": "Serveur de production principal"
    }
  ],
  "paths": {
...
    "/api/webhooks/stripe": {
      "post": {
        "summary": "Webhook de notification de paiement Stripe",
        "tags": [
          "Webhooks"
        ],
        "description": "Point d'entrée de callback appelé par Stripe pour notifier des événements de transaction (succès ou échec).",
        "responses": {
          "200": {
            "description": "Événement reçu et traité avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "received": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Échec du traitement ou signature invalide."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Authentification par cookie NextAuth.js"
      },
      "GlovoSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Glovo-Signature",
        "description": "Signature HMAC cryptographique pour valider les requêtes de Glovo."
      }
    },
...
  },
...
}